Forms: Checks and Radios: Customizing FormCheck rendering
When you need tighter control, or want to customize how the FormCheck component renders, it may better to use its constituent parts directly.

By provided children to the FormCheck you can forgo the default rendering and handle it yourself. (You can still provide an id to the FormCheck or FormGroup and have it propagate to the label and input).