Layout: Form groups
The FormGroup component is the easiest way to add some structure to forms. It provides a flexible container for grouping of labels, controls, optional help text, and form validation messaging. Use it with fieldsets, divs, or nearly any other element.

You also add the controlId prop to accessibly wire the nested label and input together via the id.