Components: Navbars: Responsive Behaviors
Use the expand prop as well as the Navbar.Toggle and Navbar.Collapse components to control when content collapses behind a button.

Set the defaultExpanded prop to make the Navbar start expanded. Set collapseOnSelect to make the Navbar collapse automatically when the user selects an item. You can also finely control the collapsing behavior by using the expanded and onToggle props.

Caution

Watch out! You need to provide a breakpoint value to expand in order for the Navbar to collapse at all.