Components: Overlay: Updating Position Dynamically
Since we can't know every time your overlay changes size, to reposition it, you need to take manual action if you want to update the position of an Overlay in response to a change.

For this, the Overlay component also injects a popper prop with a scheduleUpdate() method that an overlay component can use to reposition itself.