Components: Placeholders
How it works

Create placeholders with the Placeholder component and a grid column prop (e.g., xs={6}) to set the width. They can replace the text inside an element or be added to an existing component via the as prop.

Additional styling is applied to PlaceholderButtons via ::before to ensure the height is respected. You may extend this pattern for other situations as needed, or add a &nbsp; within the element to reflect the height when actual text is rendered in its place.